requirements.txt
docxtpl
openpyxl
python-docx
